requirements.txt
fastapi[all]
requests
SQLAlchemy
pydantic
python-dotenv
uvicorn
mysqlclient
cloudinary
bcrypt
jwt
pillow